Rename TestClientLevelContext -> TestServerConnection, TestServerConnection -> TestDedicatedServerConnection
