+++ fabric-client-gametest-api-v1/src/client/java/net/fabricmc/fabric/api/client/gametest/v1/context/TestDedicatedServerContext.java
@@ -37,7 +37,7 @@ public interface TestDedicatedServerContext extends TestServerContext, AutoClose
 	 *
 	 * @return The connection handle to the dedicated server
 	 */
-	TestServerConnection connect();
+	TestDedicatedServerConnection connect();
 
 	/**
 	 * Stops the dedicated server.
