+++ fabric-client-gametest-api-v1/src/client/java/net/fabricmc/fabric/impl/client/gametest/context/TestSingleplayerContextImpl.java
@@ -23,7 +23,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.server.MinecraftServer;
 
 import net.fabricmc.fabric.api.client.gametest.v1.context.ClientGameTestContext;
-import net.fabricmc.fabric.api.client.gametest.v1.context.TestClientLevelContext;
+import net.fabricmc.fabric.api.client.gametest.v1.context.TestServerConnection;
 import net.fabricmc.fabric.api.client.gametest.v1.context.TestServerContext;
 import net.fabricmc.fabric.api.client.gametest.v1.context.TestSingleplayerContext;
 import net.fabricmc.fabric.api.client.gametest.v1.world.TestWorldSave;
@@ -33,13 +33,13 @@ public class TestSingleplayerContextImpl implements TestSingleplayerContext {
 	private final ClientGameTestContext context;
 	private final TestWorldSave worldSave;
 	private final TestServerContextImpl server;
-	private final TestClientLevelContext clientLevel;
+	private final TestServerConnection connection;
 
 	public TestSingleplayerContextImpl(ClientGameTestContext context, TestWorldSave worldSave, MinecraftServer server) {
 		this.context = context;
 		this.worldSave = worldSave;
 		this.server = new TestServerContextImpl(server);
-		this.clientLevel = new TestClientLevelContextImpl(context, this.server);
+		this.connection = new TestServerConnectionImpl(context, this.server);
 	}
 
 	@Override
@@ -48,8 +48,8 @@ public class TestSingleplayerContextImpl implements TestSingleplayerContext {
 	}
 
 	@Override
-	public TestClientLevelContext getClientLevel() {
-		return clientLevel;
+	public TestServerConnection getConnection() {
+		return connection;
 	}
 
 	@Override
