+++ fabric-data-attachment-api-v1/src/testmodClient/java/net/fabricmc/fabric/test/attachment/client/gametest/PersistenceGametest.java
@@ -69,7 +69,7 @@ public class PersistenceGametest implements FabricClientGameTest {
 				.adjustSettings(worldCreator -> worldCreator.setGameMode(WorldCreationUiState.SelectedGameMode.CREATIVE))
 				.create()) {
 			save = spContext.getWorldSave();
-			spContext.getClientLevel().waitForChunksDownload();
+			spContext.getConnection().waitForChunksDownload();
 
 			spContext.getServer().runOnServer(server -> {
 				ServerLevel overworld = server.overworld();
@@ -99,7 +99,7 @@ public class PersistenceGametest implements FabricClientGameTest {
 
 		// second launch
 		try (TestSingleplayerContext spContext = save.open()) {
-			spContext.getClientLevel().waitForChunksDownload();
+			spContext.getConnection().waitForChunksDownload();
 
 			LOGGER.info("Testing persistent attachments");
 			spContext.getServer().runOnServer(server -> {
@@ -131,7 +131,7 @@ public class PersistenceGametest implements FabricClientGameTest {
 			LOGGER.info("Testing ProtoChunk transfer");
 			// load far chunk
 			spContext.getServer().runCommand("tp @p 4800 ~ 0");
-			spContext.getClientLevel().waitForChunksDownload();
+			spContext.getConnection().waitForChunksDownload();
 
 			spContext.getServer().runOnServer(server -> {
 				LevelChunk farChunk = server.overworld().getChunk(FAR_CHUNK_POS.x(), FAR_CHUNK_POS.z());
