+++ fabric-rendering-v1/src/testmodClient/java/net/fabricmc/fabric/test/rendering/client/LevelRenderEventsTests.java
@@ -123,7 +123,7 @@ public class LevelRenderEventsTests implements ClientModInitializer, FabricClien
 			singleplayer.getServer().runCommand("/setblock 0 99 -3 minecraft:stone");
 			singleplayer.getServer().runCommand("/tp @a 0 100 -3");
 			singleplayer.getServer().runCommand("/setblock 0 101 0 minecraft:diamond_block");
-			singleplayer.getClientLevel().waitForChunksRender();
+			singleplayer.getConnection().waitForChunksRender();
 			context.waitTicks(10);
 			context.assertScreenshotEquals(TestScreenshotComparisonOptions.of("level_render_events_block_outline_and_after_translucent").withRegion(356, 98, 142, 238).save());
 		}
